✨ 기능 추가 : MemberService 인터페이스 생성
